implement the `'` and `execute` words
This branch adds an implementation of the `'` word, which takes the name
of a word and places that word's address on the data stack, and the
`execute` word, which pops the first item on the data stack and treats
it as a pointer to a word which is then pushed to the call stack.

The `execute` function does *not* perform any checks to ensure that the
word from the data stack points to a valid word definition. Because
YOLO.

```
running 1 test
: hello ." hello, world!" ;
=> ok.
: goodbye ." goodbye, world!" ;
=> ok.
' goodbye
=> ok.
' hello
=> ok.
execute
=> hello, world!ok.
execute
=> goodbye, world!ok.
test test::execute ... ok
```