Change around how dropping an OwnedDict works a little
Prior to this change, miri complained about leaked memory.
As far as I can tell, DictionaryInner *had* a drop impl,
but it was never actually called, as the OwnedDict held
a Dictionary instead.

This moves DictionaryInner to be a private type only used
as a sort of "layout shim".

Miri tests (only on the dictionary module) pass again
after this change.