allow `DropDict::drop_dict` to take `&self`
MnemOS does not currently have an ambiently-available global heap.
Instead, allocation and deallocation is performed through an `&'static
Heap` reference. The current `DropDict` interface doesn't support this,
as it takes only the pointer and layout to deallocate. This was so that
we don't need to make the various dictionary types (and by extension,
everything) over a `D: DropDict` generic parameter.

In order to support `mnemos-alloc`'s heap, this change modifies
`DropDict` to also take an `&self` ref. Rather than changing the
dictionary types to hold a `D: DropDict` (and thus make *everything*
generic over an additional type parameter), we instead change the type
of the `deallocate` function pointer in the `Dictionary` struct to a
`fn(*const (), NonNull<u8>, Layout)`, and add a new `deallocator` field,
which is a `*const ()`. The `OwnedDict::new()` function can now take a
`&'static D` where `D: DropDict`, and store that in the `deallocator`
field as a type erased `*const ()`. The `deallocate` function pointer
now points to a generic `erased_deallocate` function, which takes the
`*const ()` and casts it back to a `D` and calls its `drop_dict`
function. This allows us to type-erase the deallocator while still
giving it a `&self` param.